+++ buildSrc/src/main/kotlin/org/jetbrains/exposed/gradle/Versions.kt
@@ -1,0 +1,22 @@
+package org.jetbrains.exposed.gradle
+
+object Versions {
+    const val kotlin = "1.3.72"
+    const val kotlinCoroutines = "1.3.5"
+
+    /** JDBC drivers **/
+    const val h2 = "1.4.199"
+    const val mariaDB = "2.6.0"
+    const val mysqlMxj = "5.0.12"
+    const val mysql51 = "5.1.49"
+    const val mysql80 = "8.0.20"
+    const val oracle12 = "12.2.0.1"
+    const val postgre = "42.2.12.jre6"
+    const val postgreNG = "0.8.4"
+    const val sqlLite3 = "3.31.1"
+    const val sqlserver = "8.2.2.jre8"
+
+    /** Spring **/
+    const val springFramework = "5.2.0.RELEASE"
+    const val springBoot = "2.2.0.RELEASE"
+}